+++ tests/test_services.py
@@ -1,21 +1,26 @@
+import pickle
+import datetime
+
+import six
+from django import forms
+from django.test import TestCase
+
+from service_objects.errors import InvalidInputsError
+from service_objects.services import ModelService
+from tests.models import CustomFooModel, FooModel
+from tests.services import (FooService, MockService, NoDbTransactionService,
+                            FooModelService)
+
 try:
     from unittest.mock import Mock, patch
 except ImportError:
     from mock import Mock, patch
 
-from unittest import TestCase
-
-import six
-from django import forms
-
-from service_objects.errors import InvalidInputsError
-from service_objects.services import ModelService
-from tests.models import FooModel
-from tests.services import FooService, MockService, NoDbTransactionService
 
 MockService.process = Mock()
+MockService.post_process = Mock()
 NoDbTransactionService.process = Mock()
-
+NoDbTransactionService.post_process = Mock()
 
 class ServiceTest(TestCase):
 
@@ -50,6 +55,16 @@ class ServiceTest(TestCase):
 
         MockService.execute({'bar': 'Hello'})
         assert mock_transaction.atomic.return_value.__enter__.called
+
+    @patch('service_objects.services.transaction')
+    def test_has_post_process_action_flag(self, mock_transaction):
+        NoDbTransactionService.execute({})
+        assert not mock_transaction.atomic.return_value.__enter__.called
+        NoDbTransactionService.process.assert_called_with()
+
+        MockService.execute({'bar': 'Hello'})
+        assert mock_transaction.atomic.return_value.__enter__.called
+        assert mock_transaction.on_commit.called_once_with(MockService.post_process)
 
 
 class ModelServiceTest(TestCase):